refactor: move restore scroll to separate hook
